+++ package.json
@@ -20,9 +20,8 @@
     "module": "node ./scripts/templates/app.js",
     "lint": "./node_modules/.bin/eslint --config ./.eslintrc.js --ext \".ts, .tsx\" src",
     "lint:fix": "./node_modules/.bin/eslint --config ./.eslintrc.js --ext \".ts, .tsx\" src --fix",
-    "build": "yarn lint && yarn build:docs && yarn build:bundles && yarn build:tabs",
-    "build:bundles": "./node_modules/.bin/tsc && rollup --config rollup.bundles.config.js",
-    "build:tabs": "./node_modules/.bin/tsc && rollup --config rollup.tabs.config.js",
+    "build": "yarn lint && yarn build:docs && yarn build:modules", 
+    "build:modules": "./node_modules/.bin/tsc && rollup --config rollup.config.js",
     "build:docs": "node ./scripts/docs/index.js",
     "serve": "http-server --cors=* -c-1 -p 8022 ./build",
     "prepare": "husky install",
