+++ rollup.config.js
@@ -52,25 +52,67 @@ const defaultConfigurations = {
 };
 
 export default (args) => {
-  let moduleBundles;
+  // Get desired modules to build
+  const definedModules = Array.from(Object.keys(modules));
 
+  let moduleBundles;
   if (args.module !== undefined) {
-    if (typeof args.module === 'string') {
-      moduleBundles = args.module.split(',');
+    if (typeof args.moduleBundles === 'string') {
+      moduleBundles = [args.module];
     } else {
-      moduleBundles = args.module;
+      moduleBundles = args.module.split(',');
     }
 
-    const unknowns = moduleBundles.filter(
-      (x) => !Object.keys(modules).includes(x)
+    const undefineds = moduleBundles.filter(
+      (val) => !definedModules.includes(val)
     );
-
-    if (unknowns.length > 0) {
-      throw new Error(`Unknown modules: ${unknowns.join(', ')}`);
+    if (undefineds.length > 0) {
+      throw new Error(`Unknown modules: ${undefineds.join(', ')}`);
     }
 
     delete args.module;
+  } else {
+    moduleBundles = definedModules;
   }
+
+  // Get desired tabs to build
+  const definedTabs = Array.from(
+    new Set(definedModules.flatMap((module) => modules[module].tabs))
+  );
+
+  let moduleTabs;
+
+  if (args.tab !== undefined) {
+    if (typeof args.tab === 'string') {
+      moduleTabs = args.tab.split(',');
+    } else {
+      moduleTabs = args.tab;
+    }
+
+    const undefineds = moduleTabs.filter((val) => !definedTabs.includes(val));
+
+    if (undefineds.length > 0) {
+      throw new Error(`Unknown tabs: ${undefineds.join(', ')}`);
+    }
+
+    delete args.tab;
+  } else {
+    moduleTabs = definedTabs;
+  }
+
+  const buildTabs = (name) => ({
+    ...defaultConfigurations,
+    input: `./src/tabs/${name}/index.tsx`,
+    output: {
+      file: `./build/tabs/${name}.js`,
+      format: 'iife',
+      globals: {
+        react: 'React',
+        'react-dom': 'ReactDom',
+      },
+    },
+    external: ['react', 'react-dom'],
+  });
 
   const buildBundle = (name) => ({
     ...defaultConfigurations,
@@ -80,10 +122,11 @@ export default (args) => {
       format: 'iife',
     },
   });
-
-  // eslint-disable-next-line no-console
   console.log(chalk.blueBright('Building bundles:'));
   moduleBundles.forEach((module) => console.log(chalk.green(module)));
 
-  return moduleBundles.map(buildBundle);
+  console.log(chalk.blueBright('Building tabs:'));
+  moduleTabs.forEach((tab) => console.log(chalk.yellow(tab)));
+
+  return [...moduleBundles.map(buildBundle), ...moduleTabs.map(buildTabs)];
 };
